backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai>=1.40.0
django-cors-headers==4.3.1
feedparser==6.0.10
requests==2.31.0
setuptools<81
celery==5.3.4
redis==5.0.1
django-celery-beat>=2.8.0

